+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - SearchBar: Added search status (min characters, loading, no results).
 - LayerSelection: Added search within current folder (requires `"filter": true"` in `config.json`).
 - Issue #1390: Added support for `.apng`, `.webp`,and `.svg` files in GFI window.
+- map3DParameter: Added new `shadowTime` configuration option to define a fixed time for shadow calculation in 3D mode.
 
 ### Changed
 - The following packages have been updated: